dateunique_function_Matlab
In remote sensing data there may be some data with 1 hour as the time precision, but we don't need such a small precision, we only need the precision in days. This function helps you to filter out the data with the same year, month and day in a certain time dataset, leaving only the data with different dates.